+++ core-indexers/tsconfig.json
@@ -3,7 +3,10 @@
     "target": "es2018",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
     "lib": ["es2021"],                                      /* Specify a set of bundled library declaration files that describe the target runtime environment. */
     "module": "commonjs",                                /* Specify what module code is generated. */
-    "rootDir": "..",    
+    "rootDir": "..", 
+    "paths": {
+      "@queryapi-runner/*": ["../runner/*"],             /* Allow imports from runner using queryapi-runner alias */
+    },
     "moduleResolution": "node",                     /* Specify how TypeScript looks up a file from a given module specifier. */
     "resolveJsonModule": true,                           /* Enable importing .json files. */
     "allowJs": true,                                     /* Allow JavaScript files to be a part of your program. Use the 'checkJS' option to get errors from these files. */
