+++ runner/src/indexer-config/index.ts
@@ -1,1 +1,2 @@
 export { default } from './indexer-config';
+export { ProvisioningConfig, LocalIndexerConfig } from './indexer-config';
