+++ runner/src/indexer/context-builder/index.ts
@@ -1,0 +1,2 @@
+export { default } from './context-builder';
+export type { ContextObject } from './context-builder';
