+++ runner/src/indexer/dml-handler/in-memory-dml-handler.test.ts
@@ -1,4 +1,4 @@
-import { type TableDefinitionNames } from '../indexer';
+import { type TableDefinitionNames } from '../../indexer';
 import InMemoryDmlHandler from './in-memory-dml-handler';
 
 const DEFAULT_ITEM_1_WITHOUT_ID = {
