+++ runner/src/indexer/dml-handler/index.ts
@@ -1,0 +1,3 @@
+export { default as DmlHandler } from './dml-handler';
+export { default as InMemoryDmlHandler } from './in-memory-dml-handler';
+export type { DmlHandlerInterface } from './dml-handler';
