+++ runner/src/indexer/index.ts
@@ -1,2 +1,3 @@
-export { default } from './indexer';
+export { default as Indexer } from './indexer';
+export { default as LocalIndexer } from './local-indexer';
 export type { TableDefinitionNames } from './indexer';
