--- runner/src/local-indexer/index.ts
@@ -1,1 +1,0 @@
-export { default } from './local-indexer';
