+++ runner/src/stream-handler/worker.ts
@@ -3,7 +3,7 @@ import { trace, type Span, context } from '@opentelemetry/api';
 import promClient from 'prom-client';
 import { Block } from '@near-lake/primitives';
 
-import Indexer from '../indexer';
+import { Indexer } from '../indexer';
 import RedisClient from '../redis-client';
 import { METRICS } from '../metrics';
 import LakeClient from '../lake-client';
@@ -14,7 +14,8 @@ import IndexerConfig from '../indexer-config';
 import parentLogger from '../logger';
 import { wrapSpan } from '../utility';
 import { type PostgresConnectionParams } from '../pg-client';
-import DmlHandler from '../dml-handler/dml-handler';
+import { DmlHandler } from '../indexer/dml-handler';
+import ContextBuilder from '../indexer/context-builder';
 
 if (isMainThread) {
   throw new Error('Worker should not be run on main thread');
@@ -102,8 +103,9 @@ async function blockQueueConsumer (workerContext: WorkerContext): Promise<void> 
   const indexerConfig: IndexerConfig = workerContext.indexerConfig;
 
   const dmlHandler: DmlHandler = new DmlHandler(workerContext.databaseConnectionParams, indexerConfig);
+  const contextBuilder: ContextBuilder = new ContextBuilder(indexerConfig, { dmlHandler });
   const indexerMeta: IndexerMeta = new IndexerMeta(indexerConfig, workerContext.databaseConnectionParams);
-  const indexer = new Indexer(indexerConfig, { dmlHandler, indexerMeta });
+  const indexer = new Indexer(indexerConfig, { contextBuilder, indexerMeta });
 
   let streamMessageId = '';
   let currBlockHeight = 0;
